% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PCR.R
\docType{data}
\name{data_PCR}
\alias{data_PCR}
\title{Database to create DNE_PCR method}
\format{
An object of class \code{data.frame} with 558 rows and 6 columns.
}
\usage{
data_PCR
}
\description{
data_PCR with 6 variables: Age,Total DNE/Total Polygon Faces, Mean value of DNE on Apex, IQR of DNE on Apex, Mean value of DNE on Convex surface, Proportion having DNE under 0.0001
}
\details{
The database was obtained by calculating the DNE on the auricular surface from European collections.
The specimens are belongs to  the Louis Lopes Collection in Lisbon, the 21st Century Identified Human Remains Collection in Coimbra, and the CAL Milano Cemetery Skeletal Collection in Milan.
}
\keyword{datasets}
